Video generation, recent files
Video generation now has a setting to get a video of only the pattern or the full preview window. It also now has a setting for frame increment, so you can set speeds between full speed and super slow.

When you save a file, it is now immediately added to the recent files list.

Data tip cursor works now.