Added new data analysis tools
I updated the Plotting and Processing files to reflect the new format. I added "create_data_analysis_tool" which can be run with inputs of exp_folder, trial_options, and a number of flags. It uses many functions which I've put in support/data_analysis_modules. I added a couple more scripts that may be useful in transititioning data analysis tools. Documentation on how to use these will be forthcoming, in the meantime, if you want to use them and have questions contact me!