Preview updated
When previewing a pattern in the designer screen, the pattern now also displays on the arena screen. Frame Forward and Frame Back buttons now allow you to tab through a pattern on the arena screen. Note that right now, the Play, Pause, and Stop buttons do not affect the arena screens, but only the in screen preview. Hitting Frame Forward or Frame Back will reset the screen to match the current pattern frame showing in the GUI.