Panel_com and preview updates
Displaying in-screen preview on arena is now an option indicated by checkbox at bottom left.

Panel_com should now generate an error log file and update it whenever a command sent to the screen is unsuccessful, with timestamp and the command in question. use this new version so we can keep track of what commands have issues.