processing bugs, conductor automatic analysis
A couple spots in the processing threw errors if the data being processed was all NaNs (meaning no fly was run). This sometimes happens in testing the software, when you don't want to waste a fly on a protocol that's not a real experiment, so fixed these spots. Also had to make one more change to update the Conductor to automatically run new analysis.